[sdk] move transaction building logic into FeatureProvider
