+++ crates/sui-core/src/unit_tests/data/entry_point_types/Move.lock
@@ -20,3 +20,8 @@ source = { local = "../../../../../sui-framework/packages/sui-framework" }
 dependencies = [
   { name = "MoveStdlib" },
 ]
+
+[move.toolchain-version]
+compiler-version = "1.16.0"
+edition = "legacy"
+flavor = "sui"
