+++ sdk/typescript/src/builder/TransactionBlock.ts
@@ -3,34 +3,20 @@
 
 import type { SerializedBcs } from '@mysten/bcs';
 import { fromB64, isSerializedBcs } from '@mysten/bcs';
-import { is, mask } from 'superstruct';
+import { mask } from 'superstruct';
 
-import { bcs } from '../bcs/index.js';
-import type { ProtocolConfig, SuiClient, SuiMoveNormalizedType } from '../client/index.js';
+import type { ProtocolConfig, SuiClient } from '../client/index.js';
 import type { Keypair, SignatureWithBytes } from '../cryptography/index.js';
-import type { SuiObjectResponse } from '../types/index.js';
-import {
-	extractMutableReference,
-	extractReference,
-	extractStructTag,
-	getObjectReference,
-	SuiObjectRef,
-} from '../types/index.js';
-import { SUI_TYPE_ARG } from '../utils/index.js';
-import { normalizeSuiAddress, normalizeSuiObjectId } from '../utils/sui-types.js';
+import { SuiObjectRef } from '../types/index.js';
+import { normalizeSuiAddress } from '../utils/sui-types.js';
+import type { FeatureProvider } from './FeatureProvider.js';
+import { DefaultFeatureProvider } from './FeatureProvider.js';
 import type { ObjectCallArg } from './Inputs.js';
-import {
-	BuilderCallArg,
-	getIdFromCallArg,
-	Inputs,
-	isMutableSharedObjectInput,
-	PureCallArg,
-} from './Inputs.js';
+import { getIdFromCallArg, Inputs } from './Inputs.js';
 import { createPure } from './pure.js';
-import { getPureSerializationType, isTxContext } from './serializer.js';
 import type { TransactionExpiration } from './TransactionBlockData.js';
 import { TransactionBlockDataBuilder } from './TransactionBlockData.js';
-import type { MoveCallTransaction, TransactionArgument, TransactionType } from './Transactions.js';
+import type { TransactionArgument, TransactionType } from './Transactions.js';
 import { TransactionBlockInput, Transactions } from './Transactions.js';
 import { create } from './utils.js';
 
@@ -95,18 +81,6 @@ function createTransactionResult(index: number): TransactionResult {
 	}) as TransactionResult;
 }
 
-function isReceivingType(normalizedType: SuiMoveNormalizedType): boolean {
-	const tag = extractStructTag(normalizedType);
-	if (tag) {
-		return (
-			tag.Struct.address === '0x2' &&
-			tag.Struct.module === 'transfer' &&
-			tag.Struct.name === 'Receiving'
-		);
-	}
-	return false;
-}
-
 function expectClient(options: BuildOptions): SuiClient {
 	if (!options.client) {
 		throw new Error(
@@ -131,17 +105,6 @@ const LIMITS = {
 } as const;
 
 type Limits = Partial<Record<keyof typeof LIMITS, number>>;
-
-// An amount of gas (in gas units) that is added to transactions as an overhead to ensure transactions do not fail.
-const GAS_SAFE_OVERHEAD = 1000n;
-
-// The maximum objects that can be fetched at once using multiGetObjects.
-const MAX_OBJECTS_PER_FETCH = 50;
-
-const chunk = <T>(arr: T[], size: number): T[][] =>
-	Array.from({ length: Math.ceil(arr.length / size) }, (_, i) =>
-		arr.slice(i * size, i * size + size),
-	);
 
 interface BuildOptions {
 	client?: SuiClient;
@@ -532,275 +495,17 @@ export class TransactionBlock {
 		return this.#blockData.getDigest();
 	}
 
-	#validate(options: BuildOptions) {
-		const maxPureArgumentSize = this.#getConfig('maxPureArgumentSize', options);
-		// Validate all inputs are the correct size:
-		this.#blockData.inputs.forEach((input, index) => {
-			if (is(input.value, PureCallArg)) {
-				if (input.value.Pure.length > maxPureArgumentSize) {
-					throw new Error(
-						`Input at index ${index} is too large, max pure input size is ${maxPureArgumentSize} bytes, got ${input.value.Pure.length} bytes`,
-					);
-				}
-			}
-		});
+	async #prepareGasPrice(featureProvider: FeatureProvider) {
+		await featureProvider.resolveFeature('txb:setGasPrice', this.#blockData);
 	}
 
-	// The current default is just picking _all_ coins we can which may not be ideal.
-	async #prepareGasPayment(options: BuildOptions) {
-		if (this.#blockData.gasConfig.payment) {
-			const maxGasObjects = this.#getConfig('maxGasObjects', options);
-			if (this.#blockData.gasConfig.payment.length > maxGasObjects) {
-				throw new Error(`Payment objects exceed maximum amount: ${maxGasObjects}`);
-			}
-		}
-
-		// Early return if the payment is already set:
-		if (options.onlyTransactionKind || this.#blockData.gasConfig.payment) {
-			return;
-		}
-
-		const gasOwner = this.#blockData.gasConfig.owner ?? this.#blockData.sender;
-
-		const coins = await expectClient(options).getCoins({
-			owner: gasOwner!,
-			coinType: SUI_TYPE_ARG,
-		});
-
-		const paymentCoins = coins.data
-			// Filter out coins that are also used as input:
-			.filter((coin) => {
-				const matchingInput = this.#blockData.inputs.find((input) => {
-					if (
-						is(input.value, BuilderCallArg) &&
-						'Object' in input.value &&
-						'ImmOrOwned' in input.value.Object
-					) {
-						return coin.coinObjectId === input.value.Object.ImmOrOwned.objectId;
-					}
-
-					return false;
-				});
-
-				return !matchingInput;
-			})
-			.slice(0, this.#getConfig('maxGasObjects', options) - 1)
-			.map((coin) => ({
-				objectId: coin.coinObjectId,
-				digest: coin.digest,
-				version: coin.version,
-			}));
-
-		if (!paymentCoins.length) {
-			throw new Error('No valid gas coins found for the transaction.');
-		}
-
-		this.setGasPayment(paymentCoins);
-	}
-
-	async #prepareGasPrice(options: BuildOptions) {
-		if (options.onlyTransactionKind || this.#blockData.gasConfig.price) {
-			return;
-		}
-
-		this.setGasPrice(await expectClient(options).getReferenceGasPrice());
-	}
-
-	async #prepareTransactions(options: BuildOptions) {
-		const { inputs, transactions } = this.#blockData;
-
-		const moveModulesToResolve: MoveCallTransaction[] = [];
-
-		// Keep track of the object references that will need to be resolved at the end of the transaction.
-		// We keep the input by-reference to avoid needing to re-resolve it:
-		const objectsToResolve: {
-			id: string;
-			input: TransactionBlockInput;
-			normalizedType?: SuiMoveNormalizedType;
-		}[] = [];
-
-		inputs.forEach((input) => {
-			if (input.type === 'object' && typeof input.value === 'string') {
-				// The input is a string that we need to resolve to an object reference:
-				objectsToResolve.push({ id: normalizeSuiAddress(input.value), input });
-				return;
-			}
-		});
-
-		transactions.forEach((transaction) => {
-			// Special case move call:
-			if (transaction.kind === 'MoveCall') {
-				// Determine if any of the arguments require encoding.
-				// - If they don't, then this is good to go.
-				// - If they do, then we need to fetch the normalized move module.
-				const needsResolution = transaction.arguments.some(
-					(arg) => arg.kind === 'Input' && !is(inputs[arg.index].value, BuilderCallArg),
-				);
-
-				if (needsResolution) {
-					moveModulesToResolve.push(transaction);
-				}
-			}
-
-			// Special handling for values that where previously encoded using the wellKnownEncoding pattern.
-			// This should only happen when transaction block data was hydrated from an old version of the SDK
-			if (transaction.kind === 'SplitCoins') {
-				transaction.amounts.forEach((amount) => {
-					if (amount.kind === 'Input') {
-						const input = inputs[amount.index];
-						if (typeof input.value !== 'object') {
-							input.value = Inputs.Pure(bcs.U64.serialize(input.value));
-						}
-					}
-				});
-			}
-
-			if (transaction.kind === 'TransferObjects') {
-				if (transaction.address.kind === 'Input') {
-					const input = inputs[transaction.address.index];
-					if (typeof input.value !== 'object') {
-						input.value = Inputs.Pure(bcs.Address.serialize(input.value));
-					}
-				}
-			}
-		});
-
-		if (moveModulesToResolve.length) {
-			await Promise.all(
-				moveModulesToResolve.map(async (moveCall) => {
-					const [packageId, moduleName, functionName] = moveCall.target.split('::');
-
-					const normalized = await expectClient(options).getNormalizedMoveFunction({
-						package: normalizeSuiObjectId(packageId),
-						module: moduleName,
-						function: functionName,
-					});
-
-					// Entry functions can have a mutable reference to an instance of the TxContext
-					// struct defined in the TxContext module as the last parameter. The caller of
-					// the function does not need to pass it in as an argument.
-					const hasTxContext =
-						normalized.parameters.length > 0 && isTxContext(normalized.parameters.at(-1)!);
-
-					const params = hasTxContext
-						? normalized.parameters.slice(0, normalized.parameters.length - 1)
-						: normalized.parameters;
-
-					if (params.length !== moveCall.arguments.length) {
-						throw new Error('Incorrect number of arguments.');
-					}
-
-					params.forEach((param, i) => {
-						const arg = moveCall.arguments[i];
-						if (arg.kind !== 'Input') return;
-						const input = inputs[arg.index];
-						// Skip if the input is already resolved
-						if (is(input.value, BuilderCallArg)) return;
-
-						const inputValue = input.value;
-
-						const serType = getPureSerializationType(param, inputValue);
-
-						if (serType) {
-							input.value = Inputs.Pure(inputValue, serType);
-							return;
-						}
-
-						const structVal = extractStructTag(param);
-						if (structVal != null || (typeof param === 'object' && 'TypeParameter' in param)) {
-							if (typeof inputValue !== 'string') {
-								throw new Error(
-									`Expect the argument to be an object id string, got ${JSON.stringify(
-										inputValue,
-										null,
-										2,
-									)}`,
-								);
-							}
-							objectsToResolve.push({
-								id: inputValue,
-								input,
-								normalizedType: param,
-							});
-							return;
-						}
-
-						throw new Error(
-							`Unknown call arg type ${JSON.stringify(param, null, 2)} for value ${JSON.stringify(
-								inputValue,
-								null,
-								2,
-							)}`,
-						);
-					});
-				}),
-			);
-		}
-
-		if (objectsToResolve.length) {
-			const dedupedIds = [...new Set(objectsToResolve.map(({ id }) => id))];
-			const objectChunks = chunk(dedupedIds, MAX_OBJECTS_PER_FETCH);
-			const objects = (
-				await Promise.all(
-					objectChunks.map((chunk) =>
-						expectClient(options).multiGetObjects({
-							ids: chunk,
-							options: { showOwner: true },
-						}),
-					),
-				)
-			).flat();
-
-			let objectsById = new Map(
-				dedupedIds.map((id, index) => {
-					return [id, objects[index]];
-				}),
-			);
-
-			const invalidObjects = Array.from(objectsById)
-				.filter(([_, obj]) => obj.error)
-				.map(([id, _]) => id);
-			if (invalidObjects.length) {
-				throw new Error(`The following input objects are invalid: ${invalidObjects.join(', ')}`);
-			}
-
-			objectsToResolve.forEach(({ id, input, normalizedType }) => {
-				const object = objectsById.get(id)!;
-				const owner = object.data?.owner;
-				const initialSharedVersion =
-					owner && typeof owner === 'object' && 'Shared' in owner
-						? owner.Shared.initial_shared_version
-						: undefined;
-
-				if (initialSharedVersion) {
-					// There could be multiple transactions that reference the same shared object.
-					// If one of them is a mutable reference or taken by value, then we should mark the input
-					// as mutable.
-					const isByValue =
-						normalizedType != null &&
-						extractMutableReference(normalizedType) == null &&
-						extractReference(normalizedType) == null;
-					const mutable =
-						isMutableSharedObjectInput(input.value) ||
-						isByValue ||
-						(normalizedType != null && extractMutableReference(normalizedType) != null);
-
-					input.value = Inputs.SharedObjectRef({
-						objectId: id,
-						initialSharedVersion,
-						mutable,
-					});
-				} else if (normalizedType && isReceivingType(normalizedType)) {
-					input.value = Inputs.ReceivingRef(getObjectReference(object)!);
-				} else {
-					input.value = Inputs.ObjectRef(getObjectReference(object as SuiObjectResponse)!);
-				}
-			});
-		}
+	async #prepareTransactions(featureProvider: FeatureProvider) {
+		await featureProvider.resolveFeature('txb:normalizeInputs', this.#blockData);
+		await featureProvider.resolveFeature('txb:resolveObjectReferences', this.#blockData);
 	}
 
 	/**
-	 * Prepare the transaction by valdiating the transaction data and resolving all inputs
+	 * Prepare the transaction by validating the transaction data and resolving all inputs
 	 * so that it can be built into bytes.
 	 */
 	async #prepare(options: BuildOptions) {
@@ -812,48 +517,27 @@ export class TransactionBlock {
 			options.protocolConfig = await options.client.getProtocolConfig();
 		}
 
-		await Promise.all([this.#prepareGasPrice(options), this.#prepareTransactions(options)]);
+		const featureProvider = new DefaultFeatureProvider([], () => expectClient(options));
+		await this.#prepareTransactions(featureProvider);
 
 		if (!options.onlyTransactionKind) {
-			await this.#prepareGasPayment(options);
+			// TODO: this was previously done in parallel with prepareTransactions, should we allow feature providers to execute in parallel?
+			await this.#prepareGasPrice(featureProvider);
 
-			if (!this.#blockData.gasConfig.budget) {
-				const dryRunResult = await expectClient(options).dryRunTransactionBlock({
-					transactionBlock: this.#blockData.build({
-						maxSizeBytes: this.#getConfig('maxTxSizeBytes', options),
-						overrides: {
-							gasConfig: {
-								budget: String(this.#getConfig('maxTxGas', options)),
-								payment: [],
-							},
-						},
-					}),
-				});
-				if (dryRunResult.effects.status.status !== 'success') {
-					throw new Error(
-						`Dry run failed, could not automatically determine a budget: ${dryRunResult.effects.status.error}`,
-						{ cause: dryRunResult },
-					);
-				}
+			await featureProvider.resolveFeature('txb:setGasBudget', this.#blockData, {
+				maxTxGas: this.#getConfig('maxTxGas', options),
+				maxTxSizeBytes: this.#getConfig('maxTxSizeBytes', options),
+			});
 
-				const safeOverhead = GAS_SAFE_OVERHEAD * BigInt(this.blockData.gasConfig.price || 1n);
-
-				const baseComputationCostWithOverhead =
-					BigInt(dryRunResult.effects.gasUsed.computationCost) + safeOverhead;
-
-				const gasBudget =
-					baseComputationCostWithOverhead +
-					BigInt(dryRunResult.effects.gasUsed.storageCost) -
-					BigInt(dryRunResult.effects.gasUsed.storageRebate);
-
-				// Set the budget to max(computation, computation + storage - rebate)
-				this.setGasBudget(
-					gasBudget > baseComputationCostWithOverhead ? gasBudget : baseComputationCostWithOverhead,
-				);
-			}
+			// TODO: this was previously done before budgeting, but it seems like it should be done after?
+			await featureProvider.resolveFeature('txb:setGasPayment', this.#blockData, {
+				maxGasObjects: this.#getConfig('maxGasObjects', options),
+			});
 		}
 
 		// Perform final validation on the transaction:
-		this.#validate(options);
+		await featureProvider.resolveFeature('txb:validate', this.#blockData, {
+			maxPureArgumentSize: this.#getConfig('maxPureArgumentSize', options),
+		});
 	}
 }
