+++ sdk/typescript/src/builder/Transactions.ts
@@ -22,6 +22,7 @@ import {
 import type { TypeTag } from '../bcs/index.js';
 import { bcs } from '../bcs/index.js';
 import { TypeTagSerializer } from '../bcs/type-tag-serializer.js';
+import { SuiMoveNormalizedType } from '../types/normalized.js';
 import { normalizeSuiObjectId } from '../utils/sui-types.js';
 import { Inputs } from './Inputs.js';
 import { create } from './utils.js';
@@ -35,12 +36,14 @@ export const TransactionBlockInput = union([
 		index: integer(),
 		value: optional(any()),
 		type: optional(literal('object')),
+		normalizedType: optional(SuiMoveNormalizedType),
 	}),
 	object({
 		kind: literal('Input'),
 		index: integer(),
 		value: optional(any()),
 		type: literal('pure'),
+		normalizedType: optional(SuiMoveNormalizedType),
 	}),
 ]);
 export type TransactionBlockInput = Infer<typeof TransactionBlockInput>;
