+++ sdk/typescript/test/e2e/data/dynamic_fields/Move.lock
@@ -20,3 +20,8 @@ source = { local = "../../../../../../crates/sui-framework/packages/sui-framewor
 dependencies = [
   { name = "MoveStdlib" },
 ]
+
+[move.toolchain-version]
+compiler-version = "1.16.0"
+edition = "legacy"
+flavor = "sui"
